requirements.txt
requests==2.31.0
gspread==6.1.2
google-auth==2.29.0
google-analytics-data==0.18.0
python-dotenv==1.0.1
